Updated Specs and Factories to work with Active Storage
Utilization of Active Storage required a few changes to the original spec.
The original spec was created for Paperclip which had a default URL when
no icon was present. This behavior allowed the original spec to pass with
the created taxonomies and taxon. This behavior is not the same with
Active Storage, therefore modifications were also made to the factories
to append an icon image. Variables replaced the static urls in
taxonomies.json.erb to account for Active Storage's unique url generation.